aws/vocab-generator-backend/requirements.txt
opea-comps==1.2
requests==2.32.3